requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
statsmodels>=0.14.0
arch>=6.0.0
scikit-learn>=1.3.0
torch>=2.0.0
pytorch-lightning>=2.0.0
pytorch-forecasting>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tqdm>=4.65.0
yfinance>=0.2.0
pyarrow>=10.0.0

